Fixed some conversion warnings.
Fixed another possible overflow problem. PAK'ing files larger than 4GB should now be possible.
Modified PAK button to better reflect what the action does. It will display "RePAK" when a PAK file is loaded and "PAK" when one is not loaded.

+++ PAKGUI/lstEvents.cpp
@@ -16,7 +16,7 @@ System::Void frmMain::lstPakContents_ItemCheck(System::Object^  sender, System::
 
 	string tmp;
 	MarshalString( lstPakContents->Items[ e->Index ]->Text, tmp );
-	long checkedsize = fileSizes[ tmp ];
+	unsigned long long checkedsize = fileSizes[ tmp ];
 
 	// Because of the way the checked items are updated in relation to the event, we have to account for a difference of 1 on two occasions.
 	// these if-statements also allow us to perform actions specific to a check or uncheck event
@@ -24,13 +24,13 @@ System::Void frmMain::lstPakContents_ItemCheck(System::Object^  sender, System::
 	{
 		++numChecked;
 		//lblPakSizeAfterPak->Text = atoi( cursize.c_str() ) + atoi( checkedsize.c_str() ) + " KB"; // Update Estimated size of PAK file after PAK'ing
-		currentEstimatedSize += std::max( checkedsize, (long) 1024 ); // calculate the new size (remember, still in bytes). also, we want to at least display 1KB of difference
+		currentEstimatedSize += std::max( checkedsize, (unsigned long long) 1024 ); // calculate the new size (remember, still in bytes). also, we want to at least display 1KB of difference
 	}
 	else // If the event was triggered by an uncheck
 	{
 		--numChecked;
 		//lblPakSizeAfterPak->Text = atoi( cursize.c_str() ) - atoi( checkedsize.c_str() ) + " KB"; // Update Estimated size of PAK file after PAK'ing
-		currentEstimatedSize -= std::max( checkedsize, (long) 1024 ); // calculate the new size (remember, still in bytes). also, we want to at least display 1KB of difference
+		currentEstimatedSize -= std::max( checkedsize, (unsigned long long) 1024 ); // calculate the new size (remember, still in bytes). also, we want to at least display 1KB of difference
 	}
 
 	// finally, determine the new units
@@ -76,7 +76,7 @@ System::Void frmMain::lstPakContents_DragDrop(System::Object^  sender, System::W
 	for each ( String ^p in s )
 	{
 		// calc file size
-		long bytes = getFileBytes( p );
+		unsigned long long bytes = getFileBytes( p );
 		if ( bytes <= 0 )
 		{
 			if ( errors )
